+++ src/main/java/ru/sstu/Mello/service/SubtaskService.java
@@ -2,6 +2,7 @@ package ru.sstu.Mello.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import ru.sstu.Mello.exception.AccessDeniedException;
 import ru.sstu.Mello.exception.ResourceNotFoundException;
 import ru.sstu.Mello.model.Subtask;
 import ru.sstu.Mello.model.Task;
@@ -9,6 +10,7 @@ import ru.sstu.Mello.model.User;
 import ru.sstu.Mello.model.dto.SubtaskRequest;
 import ru.sstu.Mello.repository.SubtaskRepository;
 import ru.sstu.Mello.repository.TaskRepository;
+import ru.sstu.Mello.repository.UserRepository;
 import ru.sstu.Mello.security.UserPrincipal;
 
 @Service
@@ -16,6 +18,8 @@ import ru.sstu.Mello.security.UserPrincipal;
 public class SubtaskService {
     private final SubtaskRepository subtaskRepository;
     private final TaskRepository taskRepository;
+    private final UserRepository userRepository;
+    private final ProjectService projectService;
 
     public Subtask getSubtask(int id) {
         return subtaskRepository.findById(id)
@@ -25,6 +29,8 @@ public class SubtaskService {
     public void addSubtask(int taskId, SubtaskRequest subtaskRequest, UserPrincipal currentUser) {
         Task task = taskRepository.findById(taskId)
                 .orElseThrow(() -> new ResourceNotFoundException("Task", "id", taskId));
+
+        projectService.checkAccess(task.getList().getProject().getId(), currentUser);
 
         Subtask subtask = Subtask.builder()
                 .title(subtaskRequest.getTitle())
@@ -37,11 +43,15 @@ public class SubtaskService {
 
     public void deleteSubtask(int subtaskId, UserPrincipal currentUser) {
         Subtask subtask = getSubtask(subtaskId);
+        projectService.checkAccess(subtask.getTask().getList().getProject().getId(), currentUser);
+
         subtaskRepository.delete(subtask);
     }
 
     public void updateStatus(int subtaskId, boolean status, UserPrincipal currentUser) {
         Subtask subtask = getSubtask(subtaskId);
+        projectService.checkAccess(subtask.getTask().getList().getProject().getId(), currentUser);
+
         subtask.setComplete(status);
         subtaskRepository.save(subtask);
     }
